kakao farmer/requirements.txt
fastapi
uvicorn
tortoise-orm
pydantic
email-validator
python-multipart
pyjwt
passlib[bcrypt]